+++ src/main/webapp/WEB-INF/views/company/join.jsp
@@ -43,7 +43,7 @@
 	</div>
 
 </div>
-<script src="/js/company.js"></script>
+<script src="/js/companyjoin.js"></script>
 <script src="/js/join.js"></script>
 <script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
 </body>
